+++ package.json
@@ -2,11 +2,13 @@
   "name": "fasttravelbe",
   "version": "1.0.0",
   "description": "Travel POC App",
-  "main": "server.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "SET NODE_ENV=production && node server",
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\""
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\"",
+    "build": "tsc -p tsconfig.json",
+    "clean": "rimraf coverage dist tmp"
   },
   "repository": {
     "type": "git",
@@ -26,6 +28,7 @@
     "@types/express": "^4.17.17",
     "@types/node": "^20.1.0",
     "nodemon": "^2.0.22",
+    "rimraf": "^5.0.0",
     "typescript": "^5.0.4"
   }
 }
